% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backward_HMM.R
\name{backward_HMM}
\alias{backward_HMM}
\title{HMM backward algorithm}
\usage{
backward_HMM(P, E, v, y, s)
}
\arguments{
\item{P}{A transition prob matrix}

\item{E}{An emission probability matrix}

\item{v}{An initial state distribution}

\item{y}{vector of observed states}

\item{s}{dim of hidden state space}
}
\description{
HMM backward algorithm
}
