% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaX2_MH.R
\name{betaX2_MH}
\alias{betaX2_MH}
\title{Metropolis-Hastings Algorithm For simulating E\link{X^2} for X~Beta(a,b)}
\usage{
betaX2_MH(N, a, b, x0, sigma)
}
\arguments{
\item{P}{A transition prob matrix initial guess}

\item{E}{An emission probability matrix initial guess}

\item{v}{An initial state distribution initial guess}

\item{y}{vector of observed states}

\item{s}{dim of hidden state space}

\item{ymax}{dim of observed state space}

\item{tol}{non-negative convergence criteria}
}
\description{
Metropolis-Hastings Algorithm For simulating E\link{X^2} for X~Beta(a,b)
}
