% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bw_HMM.R
\name{bw_HMM}
\alias{bw_HMM}
\title{Baum-Welch Algorithm for HMM}
\usage{
bw_HMM(P, E, v, y, s, ymax, tol)
}
\arguments{
\item{P}{A transition prob matrix initial guess}

\item{E}{An emission probability matrix initial guess}

\item{v}{An initial state distribution initial guess}

\item{y}{vector of observed states}

\item{s}{dim of hidden state space}

\item{ymax}{dim of observed state space}

\item{tol}{non-negative convergence criteria}
}
\description{
Baum-Welch Algorithm for HMM
}
