% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_blood.R
\name{em_blood}
\alias{em_blood}
\title{Use EM algorithm for ABO blood data}
\usage{
em_blood(n, p0, tol)
}
\arguments{
\item{n}{A 4 by 1 vector of blood type counts. n = (nA,nB,nAB,nO)}

\item{tol}{A non-negative error tolerance}

\item{p}{A 3 by 1 vector of allele probabilities. p = (pA,pB,pO)}
}
\description{
Use EM algorithm for ABO blood data
}
