% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_HMM.R
\name{forward_HMM}
\alias{forward_HMM}
\title{HMM forward algorithm}
\usage{
forward_HMM(P, E, v, y, s)
}
\arguments{
\item{P}{A transition prob matrix}

\item{E}{An emission probability matrix}

\item{v}{An initial state distribution}

\item{y}{vector of observed states}

\item{s}{dim of hidden state space}
}
\description{
HMM forward algorithm
}
