% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_logistic_regression.R
\name{gd_logistic_regression}
\alias{gd_logistic_regression}
\title{Solve logistic regression problem via gradient descent/ascent}
\usage{
gd_logistic_regression(X, y, alpha, tol)
}
\arguments{
\item{X}{A n by d matrix}

\item{y}{A n by 1 column vector}

\item{alpha}{A positive integer learning rate}

\item{tol}{A non-negative error tolerance}
}
\description{
Solve logistic regression problem via gradient descent/ascent
}
