% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr_logistic_regression.R
\name{nr_logistic_regression}
\alias{nr_logistic_regression}
\title{Solve logistic regression problem via Newton-Raphson algorithm}
\usage{
nr_logistic_regression(X, y, tol)
}
\arguments{
\item{X}{A n by d matrix}

\item{y}{A n by 1 column vector}

\item{tol}{A non-negative error tolerance}
}
\description{
Solve logistic regression problem via Newton-Raphson algorithm
}
