% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OLS_QR.R
\name{OLS_QR}
\alias{OLS_QR}
\title{Solve OLS problem using QR decomposition}
\usage{
OLS_QR(X, y)
}
\arguments{
\item{X}{A n by d matrix}

\item{y}{A n by 1 column vector}
}
\description{
Solve OLS problem using QR decomposition
}
