% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OLS_SVD.R
\name{OLS_SVD}
\alias{OLS_SVD}
\title{Solve OLS problem using SVD decomposition}
\usage{
OLS_SVD(X, y)
}
\arguments{
\item{X}{A n by d matrix}

\item{y}{A n by 1 column vector}
}
\description{
Solve OLS problem using SVD decomposition
}
