% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_MVN.R
\name{rand_MVN}
\alias{rand_MVN}
\title{Simulate N realizations of X ~ MVN(mu,Sigma)}
\usage{
rand_MVN(mu, Sigma, N)
}
\arguments{
\item{mu}{A n by 1 column vector}

\item{Sigma}{A n by n matrix}

\item{N}{An integer}
}
\description{
Simulate N realizations of X ~ MVN(mu,Sigma)
}
