% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_HMM.R
\name{sim_HMM}
\alias{sim_HMM}
\title{Simulate HMM}
\usage{
sim_HMM(P, E, V, n)
}
\arguments{
\item{P}{A transition prob matrix}

\item{E}{An emission probability matrix}

\item{n}{length of time series (# of steps)}

\item{v}{An initial state distribution}
}
\description{
Simulate HMM
}
