% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaX2_MH.R
\name{betaX2_MH}
\alias{betaX2_MH}
\title{Metropolis-Hastings Algorithm For simulating E(X^2) for X~Beta(a,b)}
\usage{
betaX2_MH(N, a, b, x0, sigma)
}
\arguments{
\item{N}{Number of MCMC steps}

\item{a}{Beta distribution parameter 1}

\item{b}{Beta distribution parameter 2}

\item{x0}{Initial guess}

\item{sigma}{Tuning parameter}
}
\description{
Metropolis-Hastings Algorithm For simulating E(X^2) for X~Beta(a,b)
}
